chore: break up sheet content segments into their own file
